requirements.txt
django == 3.1
pillow 
django-crispy-forms
django-rest-framework
